+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - Upgrade `org.jetbrains.changelog` to `1.1.1`
   - Upgrade `org.jlleitschuh.gradle.ktlint` to `10.0.0`
 - Update `changelog` Gradle plugin configuration
+- Migrate to GLES 2.0 with [Glimpse][https://glimpse.graphics/]
 
 ## [0.2.1]
 ### Added